build error fixed - 'Error:java: javacTask: source release 1.8 requires target release 1.8'
