+++ README.md
@@ -2,7 +2,7 @@ Lab First Projects
 
 Client-Server Project
 
-[![Build Status](https://travis-ci.org/Astel/Payments.svg?branch=master)](https://travis-ci.org/Astel/Payments)
+[![Build Status](https://travis-ci.org/Astel/com.epam.lab.payments.model.Payments.svg?branch=master)](https://travis-ci.org/Astel/com.epam.lab.payments.model.Payments)
 
 Описание задания: 2. Система Платежи.
 Клиент имеет одну или несколько Кредитных Карт, каждая изкоторых соответствует некоторому Счету в системе платежей.
