+++ src/main/java/com/epam/lab/payments/model/Account.java
@@ -1,4 +1,4 @@
-package epam.lab.payments.model;
+package com.epam.lab.payments.model;
 
 import lombok.Data;
