+++ src/main/java/com/epam/lab/payments/model/Card.java
@@ -1,10 +1,11 @@
-package epam.lab.payments.model;
+package com.epam.lab.payments.model;
 
 import lombok.Data;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import java.time.LocalDate;
 
 /**
  * Created by Александр on 03.11.2017.
@@ -12,12 +13,12 @@ import javax.persistence.Id;
 
 @Entity
 @Data
-public class User {
+public class Card {
     @Id
     @GeneratedValue
     private Long id;
 
-    private String name;
-
-    private String email;
+    private String cardholderName;
+    private LocalDate expiration;
+    private Integer securityCode;
 }
