+++ src/main/java/com/epam/lab/payments/model/OrderRepository.java
@@ -1,9 +1,9 @@
-package epam.lab.payments.model;
+package com.epam.lab.payments.model;
 
 import org.springframework.data.repository.Repository;
 
 /**
  * Created by Александр on 03.11.2017.
  */
-public interface CardRepository extends Repository<Card, Long> {
+public interface OrderRepository extends Repository<Order, Long> {
 }
