+++ src/main/java/com/epam/lab/payments/model/Payments.java
@@ -1,3 +1,5 @@
+package com.epam.lab.payments.model;
+
 /**
  * Created by Aleksandr_Goloshchap on 11/3/2017.
  */
